+++ src/components/CreateListModal/GameCard/index.tsx
