+++ src/components/CreateListModal/GameCard/styles.ts
