+++ src/components/GameCard/styles.ts
