--- src/components/Home/index.tsx
@@ -1,13 +1,0 @@
-import { FeaturedGames } from './FeaturedGames';
-import { Footer } from './Footer';
-import { MainSlider } from './MainSlider';
-
-export function HomeLayout() {
-  return (
-    <>
-      <MainSlider />
-      <FeaturedGames />
-      <Footer />
-    </>
-  );
-}
