+++ src/components/MainSlider/index.tsx
