--- src/pages/lists.tsx
@@ -1,5 +1,0 @@
-import { ListsLayout } from '../components/Lists';
-
-export default function Lists() {
-  return <ListsLayout />;
-}
