--- src/pages/login.tsx
@@ -1,5 +1,0 @@
-import { LoginLayout } from '../components/Login';
-
-export default function Login() {
-  return <LoginLayout />;
-}
