--- src/pages/register.tsx
@@ -1,5 +1,0 @@
-import { RegisterLayout } from '../components/Register';
-
-export default function Register() {
-  return <RegisterLayout />;
-}
